Split Claude sanitize helpers to satisfy refactor line gate
